refactor: update import paths for types and simplify method signatures across processors; remove unused types
